fix: 恢復 main 分支狀態，僅保留 X-axis 使用 Redis 時間戳
恢復內容：
- 移除 Page Visibility API（分頁切換處理）
- 移除 setupVisibilityChangeHandler() 和 reloadRecentData() 方法
- 後端 history API 恢復使用 int() 而非 float()

保留功能：
✅ X-axis 使用 Redis 數據時間戳（不使用系統時間）
   - 有數據時：value.max - 100 * 1000
   - 無數據時：'dataMin' 和 'dataMax'（不使用 new Date().getTime()）

其他更新：
- 更新資料庫 dump 檔案 (db/bim-init.sql)

現在的狀態：
- 與 main 分支一致，僅 X-axis 配置不同
- 100 秒數據窗口
- 不預載歷史資料
- 不處理分頁切換問題

+++ backend/apps/sensors/views.py
@@ -114,8 +114,8 @@ class SensorViewSet(viewsets.ModelViewSet):
         """取得歷史數據"""
         sensor = self.get_object()
 
-        # 時間範圍（支援小數，如 0.05 小時 = 3 分鐘）
-        hours = float(request.query_params.get('hours', 24))
+        # 時間範圍
+        hours = int(request.query_params.get('hours', 24))
         start_time = timezone.now() - timedelta(hours=hours)
 
         logs = SensorDataLog.objects.filter(
